requirements.txt
discord.py==2.3.2
python-dotenv
ezdxf
reportlab
Pillow
opencv-python-headless
numpy
